requirements.txt
# Flask Web Framework
Flask==2.3.3
Werkzeug==2.3.7

# Rate limiting for Flask
Flask-Limiter==3.5.0

# Environment variable management
python-dotenv==1.0.0

# HTTP requests library (used by whatsapp-python)
requests==2.31.0

# HTML parsing (used by whatsapp-python)
beautifulsoup4==4.12.2

# Async HTTP client (used by whatsapp-python)
aiohttp==3.8.6

# FastAPI (used by whatsapp-python for webhook handling)
fastapi==0.103.2
uvicorn==0.23.2

# Date and time utilities
python-dateutil==2.8.2

# Logging utilities
colorlog==6.7.0

# Development and testing dependencies
pytest==7.4.2
pytest-flask==1.3.0
pytest-cov==4.1.0

# Code formatting and linting
black==23.9.1
flake8==6.1.0

# Security scanning
bandit==1.7.5

# Production WSGI server
gunicorn==21.2.0